Fix temporary gcp default config
Before creating the final kube config
we create a temporary token which requires a default configuration.
This default configuration can either be a command path
or default gcloud config (which is why it worked locally
and on alpha environment, but not in minikube)

Since we don't actually use this feature because we use the SDK
to retrieve the temporary token, we can use fake command path.