Remove unnecessary admin membership check from user delete
Until now, when a user unregistered from the application
we checked if there are other admins in the organization
and rejected deleting the user if not.
We did this to avoid users locking themselves out from an organization.

Since organization membership is now controlled by an upstream source
we can remove this restriction.