fix(AWS): added unique client request token
At multiple workflows/activities cluster ID or generated UUID or nothing
was used. I replaced these instances with workflow ID to provide an ID that
is unique among different "requests", but are the same between activity
retries so AWS can handle same-stack requests idempotently.
ActivityID is not used, because it would be different for the same
same-stack operation between activity retries.