refactor(CF): put client behind interface/factory
This is required to properly mock activities in tests.
Unfortunately there was a very deep direct usage of
*cloudformation.CloudFormation object which could be substituted for
an AWS *session.Session + CloudFormation interface, but the cascadic
changes affected almost allactivities.
In return, activites are now much easier to be mocked and/or tested
due to one of their frequently used direct dependency is replaced
with a mockable interface which is even set up to be autogenerated.