feat(EKS/LT): made NodeVolumeSize configurable
Updated the EKS launch template for nodepools to propagate the
specified `NodeVolumeSize` to the node resource launch template's
volume size property.

In case of a 0 or unspecified value the `NodeVolumeSize` defaults
to `NodeVolumeSizeAuto` state which triggers an `AWS:NoValue` for
the volume size property which means it defaults to the AMI size.