feat(ASG/NPUpdate): added NodeVolumeSize parameter
I searched the node pool template name for usage and came up with 3
usages: `CreateASGActivity`, `UpdateASGActivity` and
`UpdateNodeGoupActivity`. I added `UsePreviousValue` logics to the
updates in this commit because explicitly set value propagation is
going to be implemented later.

The `CreateASGActivity` is extended with the `NodeVolumeSize` in a
later commit due to that change being quite extensive.