generalize node ready signal
The pke-ready signal was only sent if the (master) node succeeded with
the installation of pke. We now accept error signals as well if the node
is known to fail to bootstrap, thus we can interrupt the workflow instad
of the hopeless wait until timeout.

The renaming of the signal makes the different versions of masters and
workers incompatible.