fix(Token): fixed cluster token validation
The cluster tokens are non-expiring, but they
were stored with an explicit expiration set to 0
explicit time (0001-01-01T00:00:00Z), because at
one point where we converted between non-pointer
to pointer expiration the address of the zero
value time was taken instead of setting expiration
to nil.

This PR fixes it short-term-wise by using explicit
nil in place of the address of the zero time.

In the long run, multiple changes are needed to be
introduced to make this more robust.