chore(IS,expiry): updated cancel error handling
With Cadence client 0.17, EntityNotExistsError was
split into itself and
WorkflowExecutionAlreadyCompletedError.

Updated the logic of the Expiry IS's cancellation
error handling to handle the new error as well and
also handle all cases intuitively.

Previously if the workflow completed already, it
was handled as successful cancellation which
makes no sense, because the cluster was deleted
anyway.
This is now handled by returning an error
explaining the workflow already completed.

Also if the workflow returned continued as new
status, previously it was also handled as
successful cancellation which again doesn't make
sense, because another expiration workflow is
supposed to be scheduled in this case.
This is now handled by returning an error
indicating the cancellation should be retried.
The new workflow should be available momentarily
and the status query should returned its state
afterwards.

Other cases (failed, terminated, timed out,
already cancelled) were handled as before, namely
returning successful cancellation result.