chore(spotguides): removed scrape wf cancellation
With Cadence client 1.17, EntityNotFoundError was
split into two, itself and
WorkflowExecutionAlreadyCompletedError.

I was about to add handling the already completed
case with an error as well for cancellation
failure, when I decided to check where this
workflow is initiated to see if the new not exists
case should also be handled and found that the
referenced `scrape-shared-spotguides` workflow
doesn't seem to exist anymore, so I decided to
remove the ineffective cancellation as well.